Clear unwanted digits from BCD before converting to a string.
We want to figure out how to display the BCD number, and it's only based
on the set of digits that are passed back via bcd_to_str().  So we grab
the first non-significand digit (used for rounding-up purposes) and then
we clear out all of the others.  This gives us a nice, clean bcd object
to convert to a string.