Support for counting open and close parentheses.
This includes adding the operator_special_type, which allows the
calculator object to know when it's got a parentheses.  Other than that,
we keep track of how many parentheses we get, and then we add as many
close parentheses at the end before we run the calculation through the
infix2postfix converter so that there is an equal number of open and
close parentheses.

It's better to fix the equation so the parentheses are balanced than to
reject the equation because the user left some close parentheses off the
end of their equation.