First pass at changing the API into bcd.c and hex.c to be table driven.
This cleans up a lot of duplicate functionality in operand.c.  We no
longer have to check the base type in every single op() function to see
if it's base10 or 16 (or possibly other bases in the future).

The next step will be to get rid of all of the public functions in bcd.c
and hex.c so that their operations are completely driven from the table.